fix: bugs in remote data source

+++ lib/feature/domain/entities/cost_entity.dart
@@ -1,6 +1,6 @@
 class CostEntity {
-  final int unlock;
-  final int minute;
+  final double unlock;
+  final double minute;
 
   CostEntity({required this.unlock, required this.minute});
 }
